+++ src/main/java/de/markusbordihn/easynpc/client/screen/configuration/main/MainConfigurationScreen.java
@@ -21,7 +21,6 @@ package de.markusbordihn.easynpc.client.screen.configuration.main;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 
-import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.EditBox;
@@ -38,6 +37,7 @@ import de.markusbordihn.easynpc.Constants;
 import de.markusbordihn.easynpc.client.screen.ScreenHelper;
 import de.markusbordihn.easynpc.client.screen.configuration.ConfigurationScreen;
 import de.markusbordihn.easynpc.dialog.DialogType;
+import de.markusbordihn.easynpc.entity.ModelPose;
 import de.markusbordihn.easynpc.menu.configuration.ConfigurationType;
 import de.markusbordihn.easynpc.menu.configuration.main.MainConfigurationMenu;
 import de.markusbordihn.easynpc.network.NetworkHandler;
@@ -51,6 +51,7 @@ public class MainConfigurationScreen extends ConfigurationScreen<MainConfigurati
   protected Button editDialogButton = null;
   protected Button editEquipmentButton = null;
   protected Button editSkinButton = null;
+  protected Button poseButton = null;
   protected Button removeEntityButton = null;
   protected Button saveNameButton = null;
   protected Button scalingButton = null;
@@ -98,6 +99,15 @@ public class MainConfigurationScreen extends ConfigurationScreen<MainConfigurati
     this.saveNameButton.active = nameValue != null && !this.formerName.equals(nameValue);
   }
 
+  private Button menuButton(int left, int top, String label, Button.OnPress onPress) {
+    return menuButton(left, top, 88, label, onPress);
+  }
+
+  private Button menuButton(int left, int top, int width, String label, Button.OnPress onPress) {
+    return new Button(left, top, width, 20,
+        Component.translatable(Constants.TEXT_CONFIG_PREFIX + label), onPress);
+  }
+
   @Override
   public void init() {
     if (this.entity == null) {
@@ -109,8 +119,6 @@ public class MainConfigurationScreen extends ConfigurationScreen<MainConfigurati
     int buttonLeftPosition = this.leftPos + 110;
     int buttonSpace = 4;
     int buttonTopPosition = this.topPos + 54;
-    int buttonWidth = 88;
-    int buttonHeight = 20;
 
     // Hide home button
     this.homeButton.visible = false;
@@ -124,81 +132,94 @@ public class MainConfigurationScreen extends ConfigurationScreen<MainConfigurati
     this.nameBox.setResponder(consumer -> this.validateName());
     this.addRenderableWidget(this.nameBox);
 
-    this.saveNameButton =
-        this.addRenderableWidget(new Button(this.leftPos + 202, this.topPos + 30, buttonWidth, 20,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "save_name"), onPress -> {
-              this.saveName();
-            }));
+    this.saveNameButton = this.addRenderableWidget(
+        menuButton(this.leftPos + 202, this.topPos + 30, "save_name", onPress -> {
+          this.saveName();
+        }));
     this.saveNameButton.active = false;
 
     // Skins Button
-    this.editSkinButton =
-        this.addRenderableWidget(new Button(this.contentLeftPos, this.topPos + 205, 100, 20,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "skin"), onPress -> {
-              SkinType skinType = this.entity.getSkinType();
-              switch (skinType) {
-                case PLAYER_SKIN:
-                case SECURE_REMOTE_URL:
-                case INSECURE_REMOTE_URL:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.PLAYER_SKIN);
-                  break;
-                case CUSTOM:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.CUSTOM_SKIN);
-                  break;
-                default:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.DEFAULT_SKIN);
-              }
-            }));
+    this.editSkinButton = this.addRenderableWidget(
+        menuButton(this.contentLeftPos, this.topPos + 205, 100, "skin", onPress -> {
+          SkinType skinType = this.entity.getSkinType();
+          switch (skinType) {
+            case PLAYER_SKIN:
+            case SECURE_REMOTE_URL:
+            case INSECURE_REMOTE_URL:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.PLAYER_SKIN);
+              break;
+            case CUSTOM:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.CUSTOM_SKIN);
+              break;
+            default:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.DEFAULT_SKIN);
+          }
+        }));
 
     // Dialog Button
     this.editDialogButton = this.addRenderableWidget(
-        new Button(buttonLeftPosition, buttonTopPosition, buttonWidth, buttonHeight,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "dialog"), onPress -> {
-              DialogType dialogType = this.entity.getDialogType();
-              switch (dialogType) {
-                case BASIC:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_DIALOG);
-                  break;
-                case YES_NO:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.YES_NO_DIALOG);
-                  break;
-                default:
-                  NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_DIALOG);
-              }
-            }));
+        menuButton(buttonLeftPosition, buttonTopPosition, "dialog", onPress -> {
+          DialogType dialogType = this.entity.getDialogType();
+          switch (dialogType) {
+            case BASIC:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_DIALOG);
+              break;
+            case YES_NO:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.YES_NO_DIALOG);
+              break;
+            default:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_DIALOG);
+          }
+        }));
 
     // Actions Button
     this.editActionButton = this.addRenderableWidget(
-        new Button(this.editDialogButton.x + this.editDialogButton.getWidth() + buttonSpace,
-            buttonTopPosition, buttonWidth, buttonHeight,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "actions"), onPress -> {
+        menuButton(this.editDialogButton.x + this.editDialogButton.getWidth() + buttonSpace,
+            buttonTopPosition, "actions", onPress -> {
               NetworkHandler.openConfiguration(uuid, ConfigurationType.BASIC_ACTION);
             }));
 
-    buttonTopPosition = buttonTopPosition + buttonHeight + buttonSpace;
+    // Move button position down
+    buttonTopPosition = buttonTopPosition + 20 + buttonSpace;
 
     // Equipment Button
     this.editEquipmentButton = this.addRenderableWidget(
-        new Button(buttonLeftPosition, buttonTopPosition, buttonWidth, buttonHeight,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "equipment"), onPress -> {
-              NetworkHandler.openConfiguration(uuid, ConfigurationType.EQUIPMENT);
-            }));
+        menuButton(buttonLeftPosition, buttonTopPosition, "equipment", onPress -> {
+          NetworkHandler.openConfiguration(uuid, ConfigurationType.EQUIPMENT);
+        }));
 
     // Scaling Button
     this.scalingButton = this.addRenderableWidget(
-        new Button(this.editEquipmentButton.x + this.editEquipmentButton.getWidth() + buttonSpace,
-            buttonTopPosition, buttonWidth, buttonHeight,
-            Component.translatable(Constants.TEXT_CONFIG_PREFIX + "scaling"), onPress -> {
+        menuButton(this.editEquipmentButton.x + this.editEquipmentButton.getWidth() + buttonSpace,
+            buttonTopPosition, "scaling", onPress -> {
               NetworkHandler.openConfiguration(uuid, ConfigurationType.SCALING);
             }));
 
+    // Move button position down
+    buttonTopPosition = buttonTopPosition + 20 + buttonSpace;
+
+    // Pose Button
+    this.poseButton = this
+        .addRenderableWidget(menuButton(buttonLeftPosition, buttonTopPosition, "pose", onPress -> {
+          ModelPose modelPose = entity.getModelPose();
+          switch (modelPose) {
+            case CUSTOM:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.CUSTOM_POSE);
+              break;
+            case DEFAULT:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.DEFAULT_POSE);
+              break;
+            default:
+              NetworkHandler.openConfiguration(uuid, ConfigurationType.DEFAULT_POSE);
+          }
+        }));
+
     // Delete Button
-    this.removeEntityButton = this.addRenderableWidget(new Button(
-        this.rightPos - 60, this.bottomPos - 30, 50, 20, Component
-            .translatable(Constants.TEXT_CONFIG_PREFIX + "delete").withStyle(ChatFormatting.RED),
-        onPress -> {
+    this.removeEntityButton = this.addRenderableWidget(
+        menuButton(this.rightPos - 60, this.bottomPos - 30, 50, "delete", onPress -> {
           deleteNPC();
         }));
+    this.removeEntityButton.setFGColor(16733525);
   }
 
   @Override