+++ src/main/java/de/markusbordihn/easynpc/menu/ModMenuTypes.java
@@ -27,6 +27,8 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
 import de.markusbordihn.easynpc.Constants;
+import de.markusbordihn.easynpc.menu.configuration.pose.CustomPoseConfigurationMenu;
+import de.markusbordihn.easynpc.menu.configuration.pose.DefaultPoseConfigurationMenu;
 import de.markusbordihn.easynpc.menu.configuration.action.BasicActionConfigurationMenu;
 import de.markusbordihn.easynpc.menu.configuration.dialog.BasicDialogConfigurationMenu;
 import de.markusbordihn.easynpc.menu.configuration.dialog.YesNoDialogConfigurationMenu;
@@ -72,6 +74,14 @@ public class ModMenuTypes {
   public static final RegistryObject<MenuType<EquipmentConfigurationMenu>> EQUIPMENT_CONFIGURATION_MENU =
       MENU_TYPES.register("equipment_configuration_menu",
           () -> IForgeMenuType.create(EquipmentConfigurationMenu::new));
+
+  // Pose Configuration
+  public static final RegistryObject<MenuType<DefaultPoseConfigurationMenu>> DEFAULT_POSE_CONFIGURATION_MENU =
+      MENU_TYPES.register("default_pose_configuration_menu",
+          () -> IForgeMenuType.create(DefaultPoseConfigurationMenu::new));
+  public static final RegistryObject<MenuType<CustomPoseConfigurationMenu>> CUSTOM_POSE_CONFIGURATION_MENU =
+      MENU_TYPES.register("custom_pose_configuration_menu",
+          () -> IForgeMenuType.create(CustomPoseConfigurationMenu::new));
 
   // Skin Configuration
   public static final RegistryObject<MenuType<CustomSkinConfigurationMenu>> CUSTOM_SKIN_CONFIGURATION_MENU =